requirements.txt
functions-framework
google-cloud-firestore
google-cloud-error-reporting
mock-firestore
pytest